Refactor strength reduce
